very."(),:;<>[]".VERY."very@ "very".unusual@strange.example.com
That address is *completely* invalid, whoever came up with that
example didn't know what they were talking about.

According to the spec (rfc2821):

      Local-part = Dot-string / Quoted-string
            ; MAY be case-sensitive

      Dot-string = Atom *("." Atom)

      Atom = 1*atext

      Quoted-string = DQUOTE *qcontent DQUOTE

Even if the spec allowed for each Dot-string to be a
Quoted-string (it doesn't), it *still* wouldn't be
valid because one of the sub-tokens has unbalanced quotes.

This is why every single validator said it was invalid :)